requirements.txt
click>=8.1.0
pydantic>=2.0.0
pyyaml>=6.0
pytest>=7.0
pytest-cov>=4.0
python-dateutil>=2.8
